+++ api/src/main/java/io/strimzi/api/kafka/model/KafkaSpec.java
@@ -36,6 +36,7 @@ public class KafkaSpec extends Spec {
     private ZookeeperClusterSpec zookeeper;
     private EntityOperatorSpec entityOperator;
     private CertificateAuthority clusterCa;
+    @SuppressWarnings("deprecation")
     private JmxTransSpec jmxTrans;
     private KafkaExporterSpec kafkaExporter;
     private CruiseControlSpec cruiseControl;
@@ -96,10 +97,9 @@ public class KafkaSpec extends Spec {
     }
 
     @Deprecated
-    @DeprecatedProperty(description = "JMXTrans is deprecated and will be removed in Strimzi 0.35.0")
+    @DeprecatedProperty(description = "JMXTrans is deprecated and related resources removed in Strimzi 0.35.0.")
     @PresentInVersions("v1alpha1-v1beta2")
-    @Description("Configuration for JmxTrans. When the property is present a JmxTrans deployment is created for gathering JMX metrics from each Kafka broker. " +
-            "For more information see https://github.com/jmxtrans/jmxtrans[JmxTrans GitHub]")
+    @Description("As of Strimzi 0.35.0, JMXTrans is not supported anymore and this option is ignored.")
     public JmxTransSpec getJmxTrans() {
         return jmxTrans;
     }