+++ CHANGELOG.md
@@ -4,6 +4,13 @@
 
 * Redesigned the `UserOperatorConfig` to make it more efficient and flexible
 * Allow multiple imagePullSecrets in the Strimzi Helm chart
+* Remove support for JMX Trans
+
+### Changes, deprecations and removals
+
+* Support for JMX Trans has been removed in Strimzi 0.35.0.
+  If you have JMX Trans enabled in your `Kafka` custom resource in the `.spec.jmxTrans` section, you should remove it.
+  If you upgrade to Strimzi 0.35.0 or newer with JMX Trans deployed / enabled in the `Kafka` custom resource, Strimzi will be automatically deleted after the upgrade.
 
 ## 0.34.0